fixed bugs involving polynomials with imaginary roots and/or roots of the same magnitude but different signs

+++ include/vector/horner.h
@@ -10,4 +10,4 @@
 /* The evaluate function takes in a polynomial and an input (double), 
 then evaluates the polynomial with the given input and returns
 the correct value using the Horner method. */
-vfloat64m1_t horner(Polynomial_t poly, vfloat64m1_t vGuesses, int guessSize);
+vfloat64m1_t horner(Polynomial_t poly, vfloat64m1_t vGuesses, size_t guessSize);
